moved MeetingConfig in meeting package
