+++ src/main/java/de/slimecloud/slimeball/config/GuildConfig.java
@@ -10,7 +10,7 @@ import de.slimecloud.slimeball.features.birthday.BirthdayConfig;
 import de.slimecloud.slimeball.features.fdmds.FdmdsConfig;
 import de.slimecloud.slimeball.features.level.GuildLevelConfig;
 import de.slimecloud.slimeball.features.moderation.AutoDeleteFlag;
-import de.slimecloud.slimeball.features.staff.MeetingConfig;
+import de.slimecloud.slimeball.features.staff.meeting.MeetingConfig;
 import de.slimecloud.slimeball.features.staff.StaffConfig;
 import de.slimecloud.slimeball.features.statistic.StatisticConfig;
 import de.slimecloud.slimeball.main.Main;
